pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.3",
    "tensorflow-probability>=0.19",
    "tensorflow>=2.14.0",
    "google-meridian",
]

[tool.uv.sources]
google-meridian = { git = "https://github.com/google/meridian.git" }
